requirements.txt
filemagic>=0.4.13
pyyaml>=3.1.0
flask>=0.12.2
flask-caching>=1.3.2
requests>=2.13.0
gunicorn>=19.7.0
elasticsearch>=5.0.0,<6.0.0
elasticsearch-dsl>=5.0.0,<6.0.0
lark-parser>=0.4.1
